added policy CRUD and generation of the pdf file

+++ backend/src/main/java/com/roman/insure_manage/insuranceProduct/InsuranceProductDto.java
@@ -1,5 +1,6 @@
 package com.roman.insure_manage.insuranceProduct;
 
+import com.roman.insure_manage.common.CoverageType;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
@@ -27,7 +28,7 @@ public class InsuranceProductDto {
     private String description;
     @NotBlank(message = "Base price is required")
     @Positive(message = "Base price must be greater than 0")
-    private BigDecimal basePrice;
+    private double basePrice;
 
     @NotBlank(message = "Coverage type is required")
     private CoverageType coverageType;
