+++ backend/src/main/java/com/roman/insure_manage/insuranceProduct/InsuranceProductEntity.java
@@ -1,5 +1,6 @@
 package com.roman.insure_manage.insuranceProduct;
 
+import com.roman.insure_manage.common.CoverageType;
 import com.roman.insure_manage.insurancePolicy.InsurancePolicyEntity;
 import com.roman.insure_manage.worker.WorkerEntity;
 import jakarta.persistence.*;
@@ -31,13 +32,16 @@ public class InsuranceProductEntity  {
     private UUID id;
     private String name;
     private String description;
-    private BigDecimal basePrice;
+    private double basePrice;
 
     @Enumerated(EnumType.STRING)
     private CoverageType coverageType;
 
+
     @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     private List<InsurancePolicyEntity> policies;
+
+
     @CreatedDate
     @Column(nullable = false, updatable = false)
     private LocalDateTime createdAt;
